Added files via upload
CISC106, Spring 2016
Lab 2
Pre-lab exercises – this section is required, but nothing is submitted for grading. Read chapter 3, and chapter 5 sections 1 – 6 and 8, in the textbook. Do as many Checkpoint and Review Questions as you can. Checkpoint answers are provided in Appendix D. Some of these questions may appear on the midterm. Be sure to complete the MyProgrammingLab exercises for these chapters by the due dates posted on Sakai.
Functions:
Open IDLE, then open a new Python editor window. In the editor window, type in the following:
  print(max(1, 19, -2, 7))
  print(len('Hello'))
Run the code and examine the results. The functions max( ) and len( ) are examples of built-in functions (i.e., they are already defined for programmers to use). In fact, print() is also a built-in function. The function max() is a built-in function that returns the maximum value from a list of values, and the function len() is a built-in function that returns the length of a string. In this lab